requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-autoslug==1.9.8
django-bootstrap4==2.3.1
django-heroku==0.3.1
django-import-export==2.5.0
django-storages==1.11.1
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
whitenoise==5.2.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0